move bot specific modules to `lrrbot`
