+++ commands/highlights.py
@@ -3,9 +3,7 @@ import time
 import irc.client
 
 from common import utils
-from lrrbot import bot
-import twitch
-import storage
+from lrrbot import bot, storage, twitch
 
 
 @bot.command("highlight (.*?)")
